fix(ui) Update copy for chrome extension health component
