+++ datahub-web-react/src/app/entity/shared/containers/profile/header/PlatformContent/PlatformContentContainer.tsx
@@ -5,23 +5,8 @@ import { useEntityData } from '../../../../EntityContext';
 import { capitalizeFirstLetterOnly } from '../../../../../../shared/textUtil';
 import { getPlatformName } from '../../../../utils';
 import PlatformContentView from './PlatformContentView';
-import { GenericEntityProperties } from '../../../../types';
-import EntityRegistry from '../../../../../EntityRegistry';
-import { EntityType } from '../../../../../../../types.generated';
 import useContentTruncation from '../../../../../../shared/useContentTruncation';
-
-export function getDisplayedEntityType(
-    entityData: GenericEntityProperties | null,
-    entityRegistry: EntityRegistry,
-    entityType: EntityType,
-) {
-    return (
-        entityData?.entityTypeOverride ||
-        capitalizeFirstLetterOnly(entityData?.subTypes?.typeNames?.[0]) ||
-        entityRegistry.getEntityName(entityType) ||
-        ''
-    );
-}
+import { getDisplayedEntityType } from '../utils';
 
 function PlatformContentContainer() {
     const { entityType, entityData } = useEntityData();
