+++ datahub-web-react/src/app/entity/shared/containers/profile/header/utils.ts
@@ -1,0 +1,17 @@
+import { EntityType } from '../../../../../../types.generated';
+import { capitalizeFirstLetterOnly } from '../../../../../shared/textUtil';
+import EntityRegistry from '../../../../EntityRegistry';
+import { GenericEntityProperties } from '../../../types';
+
+export function getDisplayedEntityType(
+    entityData: GenericEntityProperties | null,
+    entityRegistry: EntityRegistry,
+    entityType: EntityType,
+) {
+    return (
+        entityData?.entityTypeOverride ||
+        capitalizeFirstLetterOnly(entityData?.subTypes?.typeNames?.[0]) ||
+        entityRegistry.getEntityName(entityType) ||
+        ''
+    );
+}
